// src/app/models/vehicleDTO.ts
export class VehicleDTO {
    fromPrice: string;
    toPrice: string;
    yearFrom: string;
    yearTo: string;
    hpFrom: string;
    hpTo: string;
    brand: string;
    model: string;
    engineType: string;
    gearboxType: string;
    categoryType: string;
    type: string;
    region: string;
    extras: Array<string>;

}